Aggiornati modelli ordini

+++ src/Ordini/OrdiniService/Models/API/OrderDTO.cs
@@ -2,18 +2,18 @@
 {
     public class OrderDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public DateTime Date { get; set; }
         public long CreationAccountId { get; set; }
 
-        public OrderDTO(int id, DateTime date, long creationAccountId)
+        public OrderDTO(long id, DateTime date, long creationAccountId)
         {
             this.Id = id;
             this.Date = date;
             this.CreationAccountId = creationAccountId;
         }
 
-        public static OrderDTO OrderDTOFactory(int id, DateTime date, long creationAccountId)
+        public static OrderDTO OrderDTOFactory(long id, DateTime date, long creationAccountId)
         {
             return new OrderDTO(id, date, creationAccountId);
         }
